+++ pkg/components/window/panel/panel.go
@@ -1,4 +1,4 @@
-package window
+package panel
 
 /*
 #cgo CFLAGS: -x objective-c
@@ -17,7 +17,7 @@ import (
 	"image"
 	"unsafe"
 
-	"github.com/kourosh/gomacui/pkg/components"
+	"github.com/kourosh/gomacui/docs/old/pkg/components"
 )
 
 // PanelStyle defines the visual style and behavior of a panel window
@@ -26,10 +26,10 @@ type PanelStyle uint
 const (
 	// PanelStyleDefault is a standard utility panel
 	PanelStyleDefault PanelStyle = iota
-	
+
 	// PanelStyleFloating is a panel that stays above regular windows
 	PanelStyleFloating
-	
+
 	// PanelStyleHUD is a panel with HUD appearance (dark, translucent)
 	PanelStyleHUD
 )
@@ -40,13 +40,13 @@ type PanelBehavior uint
 const (
 	// PanelBehaviorNormal allows the panel to be hidden when the app is inactive
 	PanelBehaviorNormal PanelBehavior = iota
-	
+
 	// PanelBehaviorCanJoinAllSpaces allows the panel to appear in all spaces
 	PanelBehaviorCanJoinAllSpaces
-	
+
 	// PanelBehaviorStaysOnActiveSpace keeps the panel in the active space
 	PanelBehaviorStaysOnActiveSpace
-	
+
 	// PanelBehaviorTransient makes the panel hide when clicking outside
 	PanelBehaviorTransient
 )
@@ -67,8 +67,8 @@ func NewPanel(title string, style PanelStyle) *Panel {
 
 	p := &Panel{
 		NativeWindow: base,
-		style:       style,
-		behavior:    PanelBehaviorNormal,
+		style:        style,
+		behavior:     PanelBehaviorNormal,
 	}
 
 	// Create native panel
@@ -87,7 +87,7 @@ func NewPanel(title string, style PanelStyle) *Panel {
 
 	// Configure panel based on style
 	p.applyStyle(style)
-	
+
 	return p
 }
 
@@ -100,7 +100,7 @@ func (p *Panel) SetStyle(style PanelStyle) {
 // SetBehavior sets the panel's behavior
 func (p *Panel) SetBehavior(behavior PanelBehavior) {
 	p.behavior = behavior
-	
+
 	var collectionBehavior C.ulong
 	switch behavior {
 	case PanelBehaviorCanJoinAllSpaces:
@@ -110,7 +110,7 @@ func (p *Panel) SetBehavior(behavior PanelBehavior) {
 	case PanelBehaviorTransient:
 		collectionBehavior = 1 << 5 // NSWindowCollectionBehaviorTransient
 	}
-	
+
 	if collectionBehavior != 0 {
 		C.setPanelCollectionBehavior(p.nsWindow, collectionBehavior)
 	}
@@ -119,7 +119,7 @@ func (p *Panel) SetBehavior(behavior PanelBehavior) {
 func (p *Panel) applyStyle(style PanelStyle) {
 	switch style {
 	case PanelStyleFloating:
-		C.setPanelLevel(p.nsWindow, 3) // NSFloatingWindowLevel
+		C.setPanelLevel(p.nsWindow, 3)    // NSFloatingWindowLevel
 		C.setPanelBehavior(p.nsWindow, 1) // NSWindowBehaviorTransient
 	case PanelStyleHUD:
 		var mask C.ulong = 1 << 12 // NSWindowStyleMaskHUDWindow
@@ -130,7 +130,7 @@ func (p *Panel) applyStyle(style PanelStyle) {
 // SetFrame overrides NativeWindow.SetFrame to maintain panel positioning
 func (p *Panel) SetFrame(frame image.Rectangle) {
 	p.NativeWindow.SetFrame(frame)
-	
+
 	// Reapply panel level if floating
 	if p.style == PanelStyleFloating {
 		C.setPanelLevel(p.nsWindow, 3) // NSFloatingWindowLevel
