+++ pkg/components/window/window/window.go
@@ -1,4 +1,4 @@
-package components
+package window
 
 import (
 	"image"
